requirements.txt
accelerate==0.27.2
astroid==3.0.3
attrs==23.2.0
autopep8==2.0.4
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
DAWG-Python==0.7.2
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
docstring-to-markdown==0.15
filelock==3.13.1
fire==0.5.0
flake8==7.0.0
fsspec==2024.2.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
googletrans==3.0.0
guidata==3.1.0
guiqwt==4.4.4
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llama_cpp_python==0.2.55
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==23.2
parso==0.8.3
peft==0.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pocketsphinx==5.0.3
psutil==5.9.8
PyAudio==0.2.14
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.0.4
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pyproject-toml==0.0.10
pyqt-tools==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
PySide2==5.13.2
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.1
PythonQwt==0.12.1
pytoolconfig==1.3.1
pyttsx3==2.90
PyYAML==6.0.1
QtPy==2.4.1
rapidfuzz==3.6.2
ratelim==0.1.6
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rope==1.13.0
rpds-py==0.18.0
safetensors==0.4.2
scipy==1.12.0
setuptools==69.1.1
shiboken2==5.13.2
shiboken6==6.6.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.5
srt==3.5.3
sympy==1.12
termcolor==2.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
vosk==0.3.45
websockets==12.0
whatthepatch==1.0.5
wheel==0.42.0
wikipedia==1.4.0
yapf==0.40.2
zipp==3.18.1